% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf2DistFFT.R
\name{cf2DistFFT}
\alias{cf2DistFFT}
\title{Evaluating CDF/PDF/QF from CF of a continous distribution F by using the FFT algorithm}
\usage{
result <- cf2DistFFT(cf)
result <- cf2DistFFT(cf, x, prob, options)
}
\arguments{
\item{cf}{function handle for the characteristic function CF}

\item{x}{vector of values from domain of the distribution F, if missing,
cf2DistFFT automatically selects vector x from the domain}

\item{prob}{vector of values from [0,1] for which the quantiles will be estimated,
default vector prob = c(0.9,0.95,0.975,0.99,0.995,0.999)}

\item{option}{list with the following (default) parameters:
\itemize{
    \item option$isCompound   = FALSE   treat the compound distributions
    \item option$N            = 2^10    set N points used by FFT
    \item option$xMin         = -Inf    set the lower limit of X
    \item option$xMax         = Inf     set the upper limit of X
    \item option$xMean        = NULL    set the MEAN value of X
    \item option$xStd         = NULL    set the STD value of X
    \item option$dt           = NULL    set grid step dt = 2*pi/xRange
    \item option$T            = NULL    set upper limit of (0,T), T = N*dt
    \item option$SixSigmaRule = 6       set the rule for computing domain
    \item option$tolDiff      = 1e-4    tol for numerical differentiation
    \item option$isPlot       = TRUE    plot the graphs of PDF/CDF

    \item option$DIST                   list with information for future evaluations.
                                        option$DIST is created automatically after first call:
    \itemize{
        \item option$DIST$xMean
        \item option$DIST$xMin
        \item option$DIST$xMax
        \item option$DIST$xRange
        \item option$DIST$cft
        \item option$DIST$N
        \item option$DIST$k
        \item option$DIST$t
        \item option$DIST$dt
        }
    }}

\item{isCompound}{treat the compound distributions, default FALSE}

\item{N}{N points used by FFT, default 2^10 (2^14 for compound distribution)}

\item{SixSigmaRule}{set the rule for computing domain, default 6}

\item{xMin}{set the lower limit of X}

\item{xMax}{set the upper limit of X}

\item{isPlot}{plot the graphs of PDF/CDF, default TRUE}
}
\value{
result - list with the following results values:
   \item{result$x}{<- x}
   \item{result$cdf}{<- cdf}
   \item{result$pdf}{<- pdf}
   \item{result$prob}{<- prob}
   \item{result$qf}{<- qf}
   \item{result$xFFT}{<- xFFT}
   \item{result$pdfFFT}{<- pdfFFT}
   \item{result$cdfFFT}{<- cdfFFT}
   \item{result$SixSigmaRule}{<- option$SixSigmaRule}
   \item{result$N}{<- option$N}
   \item{result$dt}{<- dt}
   \item{result$T}{<- t[(length(t))]}
   \item{result$PrecisionCrit}{<- PrecisionCrit}
   \item{result$myPrecisionCrit}{<- options$crit}
   \item{result$isPrecisionOK}{<- isPrecisionOK}
   \item{result$xMean}{<- xMean}
   \item{result$xSTD}{<- xStd}
   \item{result$xMin}{<- xMin}
   \item{result$xMAx}{<- xMax}
   \item{result$cf}{<- cf}
   \item{result$option}{<- option}
   \item{result$tictoc}{<- proces time}
}
\description{
\code{cf2DistFFT(cf, x, prob, options)} evaluates the CDF/PDF/QF (Cumulative Distribution Function,
Probability Density Function, Quantile Function) from the Characteristic Function CF
of a (continuous) distribution F by using the Fast Fourier Transform (FFT) algorithm.

The algorithm cf2DistFFT evaluates the approximate values CDF(x), PDF(x),
and/or the quantiles QF(prob) for given x and prob, by interpolation from
the PDF-estimate computed by the numerical inversion of the given
characteristic function CF by using the FFT algorithm.
}
\note{
The outputs of the algorithm cf2DistFFT are approximate values! The
precission of the presented results depends on several different
factors:
\itemize{
    \item  application of the FFT algorithm for numerical inversion of the CF
    \item selected number of points used by the FFT algorithm (by default
   option$N = 2^10),
    \item estimated/calculated domain [A,B] of the distribution F, used with the
   FFT algorithm. Optimally, [A,B] covers large part of the
   distribution domain, say more than 99%. However, the default
   automatic procedure for selection of the domain [A,B] may fail. It
   is based on the 'SixSigmaRule': \cr
   A = MEAN - SixSigmaRule * STD, and B = MEAN + SixSigmaRule * STD. \cr
   Alternatively, change the option$SixSigmaRule to different value, say 12,
   or use the option$xMin and option$xMax to set manually the values of A and B.
   }
}
\references{
\enumerate{
    \item WITKOVSKY, V.: On the exact computation of the density and of
    the quantiles of linear combinations of t and F random
    variables. Journal of Statistical Planning and Inference 94
    (2001), 113.
    \item WITKOVSKY, V.: Matlab algorithm TDIST: The distribution of a
    linear combination of Students t random variables. In COMPSTAT
    2004 Symposium (2004), J. Antoch, Ed., Physica-Verlag/Springer
    2004, Heidelberg, Germany, pp. 19952002.
    \item WITKOVSKY, V.: WIMMER,G., DUBY, T. Logarithmic Lambert W x F
    random variables for the family of chi-squared distributions
    and their applications. Statistics & Probability Letters 96
    (2015), 223231.
    \item WITKOVSKY V. (2016). Numerical inversion of a characteristic
    function: An alternative tool to form the probability distribution of
    output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
    \item WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
    distribution based on numerical inversion of the compound empirical
    characteristic function of frequency and severity. Preprint submitted
    to Insurance: Mathematics and Economics.
    \item WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
    distribution based on numerical inversion of the compound empirical
    characteristic function of frequency and severity. Preprint submitted
    to Insurance: Mathematics and Economics.
    \item DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM for
    computing distributions of collective risk models. Preprint submitted
    to Journal of Statistical Software.
    }
}
\seealso{


Other Algorithm: \code{\link{cf2DistGP}}
}

