% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfN_Poisson.R
\name{cfN_Poisson}
\alias{cfN_Poisson}
\title{Characteristic function of Poisson distribution}
\usage{
cfN_Poisson(t, lambda)
cfN_Poisson(t, lambda, cfX)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{lambda}{rate, lambda > 0, default value lambda = 1}

\item{cfX}{function}
}
\value{
characteristic function cf(t) of the Poisson distribution with rate lambda
}
\description{
cfN_Poisson(t, lambda) evaluates the characteristic function cf(t) of the
Poisson distribution with the rate parameter lambda > 0, i.e.
cfN_Poisson(t, lambda) = exp(lambda*(exp(1i*t)-1))

cfN_Poisson(t, lambda, cfX) evaluates the compound characteristic function
cf(t) = cfN_Poisson(-1i*log(cfX(t)), lambda), where cfX is function
handle of the characteristic function cfX(t) of a continuous distribution
and/or random variable X.

Note that such CF is characteristic function of the compound distribution,
i.e. distribution of the random variable Y = X_1 + ... + X_N, where X_i ~ F_X
are i.i.d. random variables with common CF cfX(t), and N ~ F_N is
independent RV with its CF given by cfN(t).
}
\examples{

}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Poisson_distribution}
\url{https://en.wikipedia.org/wiki/Compound_Poisson_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Discrete Probability Distribution: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_PolyaEggenberger}}
}

