% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_ChiSquared.R
\name{cfX_ChiSquared}
\alias{cfX_ChiSquared}
\title{Characteristic function of Chi-Squared distribution}
\usage{
cfX_ChiSquared(t, df)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{df}{degrees of freedom, df > 0}
}
\value{
characteristic function cf(t) of the CHI-SUQARED distribution
}
\description{
cfX_ChiSquared(t,df) evaluates the characteristic function cf(t) of
the CHI-SUQARED distribution with the parameter df (degrees of freedom, df > 0),
i.e.
cfX_ChiSquared(t, df) = (1 - 2it )^(-df/2) = cfX_GAMMA(t, df/2, 1/2)
}
\examples{
## EXAMPLE1 (CF of the ChiSquared distribution with df = 1)
 df <- 1
 t <- seq(-50, 50, length.out = 501)
 plotGraf(function(t) cfX_ChiSquared(t, df), t, title = "CF of the Chi-squared distribution with df = 1")

## EXAMPLE2 (PDF/CDF of the ChiSquared distribution with df = 3)
 df <- 3
 prob <- c(0.9, 0.95, 0.99)
 cf <- function(t) cfX_ChiSquared(t, df)
 x <- seq(-0, 15, length.out = 101)
 result <- cf2DistGP(cf, x, prob, xMin = 0, xMax = 22, N = 2^14)

## EXAMPLE3 (PDF/CDF of the compound Binomial-ChiSquared distribution)
 n <- 25
 p <- 0.3
 df <- 3
 prob <- c(0.9, 0.95, 0.99)
 cfX <- function(t) cfX_ChiSquared(t, df)
 cf <- function(t) cfN_Binomial(t, n, p, cfX)
 x <- seq(-0, 80, length.out = 501)
 result <- cf2DistGP(cf, x, prob, isCompound = TRUE)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Chi-squared_distribution}

Other Characteristic Function: \code{\link{cfC_vonMises}},
  \code{\link{cfE_Empirical}}, \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Beta}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

