% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Exponential.R
\name{cfX_Exponential}
\alias{cfX_Exponential}
\title{Characteristic function of Exponential distribution}
\usage{
cfX_Exponential(t, lambda)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{lambda}{rate, lambda > 0, default value lambda = 1}
}
\value{
characteristic function cf(t) of the Exponential distribution
}
\description{
cfX_Exponential(t, lambda) evaluates the characteristic function cf(t) of
the Exponential distribution with the parameter lambda (rate, lambda > 0)

cfX_Exponential(t, lambda) = lambda / (lambda - it)
}
\examples{

}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Exponential_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

