% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Gamma.R
\name{cfX_Gamma}
\alias{cfX_Gamma}
\title{Characteristic function of Gamma distribution}
\usage{
cfX_Gamma(t, alpha, beta)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{alpha}{shape, alpha > 0, default value alpha = 1}

\item{beta}{rate > 0, default value beta = 1}
}
\value{
characteristic function cf(t) of the Gamma distribution
}
\description{
cfX_Gamma(t, alpha, beta) evaluates the characteristic function cf(t) of
the Gamma distribution with the parameters alpha (shape, alpha > 0) and
beta (rate, beta > 0), i.e.

cfX_Gamma(t, alpha, beta) = (1 - it/beta)^(-alpha)
}
\examples{
## EXAMPLE1 (CF of the Gamma distribution with alpha = 2, beta = 2)
 alpha <- 2
 beta <- 2
 t <- seq(-20, 20, length.out = 501)
 plotGraf(function(t) cfX_Gamma(t, alpha, beta), t, title = "CF of the Gamma distribution with alpha = 2, beta = 2")

## EXAMPLE2 (PDF/CDF of the Gamma distribution with alpha = 2, beta = 2)
 alpha <- 2
 beta <- 2
 cf <- function(t) cfX_Gamma(t, alpha, beta)
 x <- seq(0, 5, length.out = 101)
 prob <- c(0.9, 0.95, 0.99)
 result <- cf2DistGP(cf, x, prob, xMin = 0, N = 2^14)

## EXAMPLE3 (PDF/CDF of the compound Binomial-Gamma distribution)
 p <- 0.3
 n <- 25
 alpha <- 2
 beta <- 2
 cfX <- function(t) cfX_Gamma(t, alpha, beta)
 cf <- function(t) cfN_Binomial(t, n, p, cfX)
 x <- seq(0, 25, length.out = 101)
 prob <- c(0.9, 0.95, 0.99)
 result <- cf2DistGP(cf, x, prob, isCompound = TRUE)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Gamma_distribution}

Other Characteristic Function: \code{\link{cfC_vonMises}},
  \code{\link{cfE_Empirical}}, \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Beta}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

