% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_InverseGamma.R
\name{cfX_InverseGamma}
\alias{cfX_InverseGamma}
\title{Characteristic function of Inverse Gamma distribution}
\usage{
cfX_InverseGamma(t, alpha, beta)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{alpha}{shape, alpha > 0, default value alpha = 1}

\item{beta}{rate > 0, default value beta = 1}
}
\value{
characteristic function cf(t) of the Inverse Gamma distribution
}
\description{
cfX_InverseGamma(t,alpha,beta) evaluates the characteristic function cf(t) of
the Inverse Gamma distribution with the parameters alpha (shape, alpha > 0) and
beta (rate, beta > 0), i.e.

cfX_InverseGamma(t, alpha, beta) = (1 - it/beta)^(-alpha)
}
\examples{
## EXAMPLE1 (CF of the Logarithmic distribution with the parameter p = 0.5)
 p <- 0.5
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p), t, title = "CF of the Logarithmic distribution with the parameter p = 0.5")

#' ## EXAMPLE2 (CF of the compound Logarithmic-Exponential distributionn)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p, cfX), t, title = "CF of the compound Logarithmic-Exponential distribution")

## EXAMPLE3 (PDF/CDF of the compound Logarithmic-Exponential distribution)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 cf <- function(t) cfN_Logarithmic(t, p, cfX)
 x <- seq(0, 3, length.out = 101)
 prob <- c(0.9, 0.95, 0.99)
 result <- cf2DistGP(cf, x, prob, isCompound = TRUE)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Inverse-gamma_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

