% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Normal.R
\name{cfX_Normal}
\alias{cfX_Normal}
\title{Characteristic function of Normal distribution}
\usage{
cfX_Normal(t, mean, variance)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{mean}{real number, mean or expextation of the distribution, default value mean = 0}

\item{variance}{real number, standard deviation, variance > 0, default value variance = 1}
}
\value{
characteristic function cf(t) of the normal distribution
}
\description{
cfX_Normal(t, mean, variance) evaluates the characteristic function cf(t) of
the Normal distribution with mean = mean and variance = variance: N(mean, variance))
cfX_Normal(t, mean, variance) = exp(imeant -1/2variance^2t^2)
}
\examples{

}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

