% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_PearsonV.R
\name{cfX_PearsonV}
\alias{cfX_PearsonV}
\title{Characteristic function of Pearson type V distribution}
\usage{
cfX_PearsonV(t, alpha, beta)
}
\arguments{
\item{t}{numerical real values (number, vector...)}

\item{alpha}{shape, alpha > 0, default value alpha = 1}

\item{beta}{scale > 0, default value beta = 1}
}
\value{
characteristic function cf(t) of the Gamma distribution
}
\description{
cfX_PearsonV(t, alpha, beta) evaluates the characteristic function cf(t) of
the Pearson type V distribution with the parameters alpha (shape, alpha > 0) and
beta (scale, beta > 0), computed for real vector argument t, i.e.

cfX_PearsonV(t, alpha, beta) = (2/gamma(alpha)) * (-1i*t/beta)^(alpha/2) * besselk(alpha,2*sqrt(-1i*t/beta)),
where besselk(a,z) denotes the modified Bessel function of the second.
}
\examples{
## EXAMPLE1 (CF of the Logarithmic distribution with the parameter p = 0.5)
 p <- 0.5
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p), t, title = "CF of the Logarithmic distribution with the parameter p = 0.5")

#' ## EXAMPLE2 (CF of the compound Logarithmic-Exponential distributionn)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p, cfX), t, title = "CF of the compound Logarithmic-Exponential distribution")

## EXAMPLE3 (PDF/CDF of the compound Logarithmic-Exponential distribution)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 cf <- function(t) cfN_Logarithmic(t, p, cfX)
 x <- seq(0, 3, length.out = 101)
 prob <- c(0.9, 0.95, 0.99)
 result <- cf2DistGP(cf, x, prob, isCompound = TRUE)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Pearson_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

