% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Rectangular.R
\name{cfX_Rectangular}
\alias{cfX_Rectangular}
\title{Characteristic function of Rectangular distribution}
\usage{
cfX_Rectangular(t, a, b)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{a}{number, default value a = -1}

\item{b}{number, default value b = 1}
}
\value{
characteristic function cf(t) of the Rectangular distribution on the interval (a, b)
}
\description{
cfX_Rectangular(t, a, b) evaluates the characteristic function cf(t) of
the Rectangular distribution on the interval (a, b)
(Rectangular distribution with mean = (a + b)/2 and variance = 1/12(b - a)^2)
cfX_Rectangular(t, a, b) = (exp(ibt) - exp(iat))/(i(b - a)t)
}
\examples{

}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Triangular}}
}

