% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Triangular.R
\name{cfX_Triangular}
\alias{cfX_Triangular}
\title{Characteristic function of Triangular distribution}
\usage{
cfX_Triangular(t, a, b, c)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{a}{number, default value a = -1}

\item{b}{number, default value b = 1}

\item{c}{number, (\eqn{a \le c \le b}), default value c = 0}
}
\value{
characteristic function cf(t) of the Triangular distribution on the interval (a, b) with mode c
}
\description{
cfX_Triangular(t, a, b, c) (\eqn{a \le c \le b}) evaluates the characteristic function cf(t) of
the Triangular distribution on the interval (a, b) with mode c
(Triangular distribution with mean = a + b + c)/3 and variance = 1/18(a^2 + b^2 + c^2 + - ab - ac - bc))
cfX_Triangula(t, a, b, c) = -2((b-c)exp(iat) - (b-a)exp(ict) + (c-a)exp(ibt))/((b-a)(c-a)(b-c)t^2)
}
\examples{
## EXAMPLE1 (CF of the Logarithmic distribution with the parameter p = 0.5)
 p <- 0.5
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p), t, title = "CF of the Logarithmic distribution with the parameter p = 0.5")

#' ## EXAMPLE2 (CF of the compound Logarithmic-Exponential distributionn)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 t <- seq(-10, 10, length.out = 501)
 plotGraf(function(t) cfN_Logarithmic(t, p, cfX), t, title = "CF of the compound Logarithmic-Exponential distribution")

## EXAMPLE3 (PDF/CDF of the compound Logarithmic-Exponential distribution)
 p <- 0.5
 lambda <- 5
 cfX <- function(t) cfX_Exponential(t, lambda)
 cf <- function(t) cfN_Logarithmic(t, p, cfX)
 x <- seq(0, 3, length.out = 101)
 prob <- c(0.9, 0.95, 0.99)
 result <- cf2DistGP(cf, x, prob, isCompound = TRUE)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Triangular_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}}, \code{\link{cfX_Rectangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_StudentT}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquaredNC}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}}, \code{\link{cfX_Rectangular}}
}

