% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolar.R
\name{plotPolar}
\alias{plotPolar}
\title{Title}
\usage{
plotPolar(angle, radius, title = "", type = "l", color = "blue")
}
\arguments{
\item{angle}{angle}

\item{radius}{amplitude}

\item{title}{an overall title for the plot}

\item{type}{type of lines, dots}

\item{color}{color}
}
\description{
Title
}
\examples{


}

