% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_ChiSquaredNC.R
\name{cfX_ChiSquaredNC}
\alias{cfX_ChiSquaredNC}
\title{Characteristic function of Noncentral Chi-Squared distribution}
\usage{
cfX_ChiSquared(t, df, npc)
}
\arguments{
\item{t}{numerical values (number, vector...)}

\item{df}{degrees of freedom}

\item{npc}{non-centrality parameter, default value = 0}
}
\value{
characteristic function cf(t) of the CHI-SUQARED distribution
}
\description{
cfX_ChiSquaredNC(t, df, ncp = 0) evaluates the characteristic function cf(t) of
the Noncentral Chi-Squared distribution with the parameter df (degrees of freedom, df > 0)
and npc (non-centrality, npc > 0), i.e.

cfX_ChiSquared(t, df, npc) = exp((i*npc*t)/(1 - 2it)) / (1 - 2it)^(df/2)
}
\examples{

}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Noncentral_chi-squared_distribution}

Other Characteristic Function: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}},
  \code{\link{cfS_Arcsine}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquared}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_Normal}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}
}

